requirements/sm52.txt
# ============================================================================
# requirements/sm52.txt
# Pip dependencies for SM52 (GTX Titan X / Maxwell) environments.
#
# DO NOT add to this file:
#   - torch, torchvision, torchaudio  (installed via conda)
#   - mkl, mkl_fft, mkl_random, mkl-service, intel-openmp  (managed by conda)
#   - diff_gaussian_rasterization, simple_knn  (built from submodules/)
#   - pip, setuptools, wheel  (conda base packages)
#
# Install with:
#   pip install -r requirements/sm52.txt
# ============================================================================

# === Core scientific ===
numpy==1.26.4
scipy==1.15.3
scikit-learn==1.7.2
scikit-image==0.25.2
pandas==2.3.3

# === 3DGS / geometry ===
open3d==0.19.0
opencv-python==4.8.1.78
plyfile==1.1.3
trimesh==4.11.2
pyquaternion==0.9.9

# === Visualization ===
matplotlib==3.10.8
plotly==6.5.2
dash==4.0.0

# === Config / CLI ===
ConfigArgParse==1.7.1
jsonc==1.1.0
PyYAML==6.0.3
tqdm==4.67.3

# === Image I/O ===
pillow==12.1.0
ImageIO==2.37.2
tifffile==2025.5.10

# === Networking / web ===
requests==2.32.5
Flask==3.1.3
Werkzeug==3.1.6
Jinja2==3.1.6
MarkupSafe==3.0.3
itsdangerous==2.2.0
blinker==1.9.0
click==8.3.1
certifi==2026.1.4
urllib3==2.6.3
idna==3.11
charset-normalizer==3.4.4

# === Jupyter / IPython ===
ipython==8.38.0
ipywidgets==8.1.8
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
widgetsnbextension==4.0.15
nbformat==5.10.4
comm==0.2.3
traitlets==5.14.3
matplotlib-inline==0.2.1

# === Utilities ===
networkx==3.4.2
joblib==1.5.3
six==1.17.0
typing_extensions==4.15.0
packaging==25.0
platformdirs==4.9.2
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.3
zipp==3.23.0
importlib_metadata==8.7.1
addict==2.4.0
retrying==1.4.2
lazy_loader==0.4
narwhals==2.17.0

# === IPython internals ===
jedi==0.19.2
parso==0.8.6
prompt_toolkit==3.0.52
Pygments==2.19.2
wcwidth==0.6.0
decorator==5.2.1
executing==2.2.1
asttokens==3.0.1
pure_eval==0.2.3
stack-data==0.6.3
pexpect==4.9.0
ptyprocess==0.7.0
exceptiongroup==1.3.1

# === JSON / schema ===
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
referencing==0.37.0
rpds-py==0.30.0
fastjsonschema==2.21.2
nest-asyncio==1.6.0

# === Plotting internals ===
contourpy==1.3.2
cycler==0.12.1
fonttools==4.61.1
kiwisolver==1.4.9
pyparsing==3.3.2
threadpoolctl==3.6.0
